# Improve `count` vectorization: replace `popcnt` implementation with vector counting