Somewhat better in a long run, some pessimization for a short one.
Before:
```
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
bm<uint8_t, Op::Count>/8021/3056        97.4 ns         95.3 ns     10000000
bm<uint8_t, Op::Count>/63/62            3.86 ns         3.13 ns    194782609
bm<uint8_t, Op::Count>/31/30            9.98 ns         8.89 ns     89600000
bm<uint8_t, Op::Count>/15/14            5.00 ns         4.97 ns    144516129
bm<uint8_t, Op::Count>/7/6              3.08 ns         3.14 ns    248888889
bm<uint16_t, Op::Count>/8021/3056        198 ns          193 ns      3733333
bm<uint16_t, Op::Count>/63/62           3.92 ns         3.75 ns    179200000
bm<uint16_t, Op::Count>/31/30           3.51 ns         3.53 ns    203636364
bm<uint16_t, Op::Count>/15/14           4.06 ns         2.13 ns    248888889
bm<uint16_t, Op::Count>/7/6             3.31 ns         1.79 ns    280000000
bm<uint32_t, Op::Count>/8021/3056        364 ns          245 ns      3446154
bm<uint32_t, Op::Count>/63/62           4.62 ns         2.20 ns    248888889
bm<uint32_t, Op::Count>/31/30           3.45 ns         2.29 ns    464592593
bm<uint32_t, Op::Count>/15/14           2.83 ns         1.50 ns    407272727
bm<uint32_t, Op::Count>/7/6             3.05 ns         2.08 ns    263529412
bm<uint64_t, Op::Count>/8021/3056        737 ns          600 ns      1120000
bm<uint64_t, Op::Count>/63/62           8.65 ns         6.25 ns    100000000
bm<uint64_t, Op::Count>/31/30           4.75 ns         3.39 ns    248888889
bm<uint64_t, Op::Count>/15/14           3.54 ns         1.79 ns    280000000
bm<uint64_t, Op::Count>/7/6             3.08 ns         2.05 ns    640000000
```

After:
```
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
bm<uint8_t, Op::Count>/8021/3056        77.5 ns         53.5 ns     27151515
bm<uint8_t, Op::Count>/63/62            4.34 ns         3.52 ns    497777778
bm<uint8_t, Op::Count>/31/30            10.1 ns         7.98 ns    172307692
bm<uint8_t, Op::Count>/15/14            5.30 ns         3.96 ns    320000000
bm<uint8_t, Op::Count>/7/6              3.50 ns         2.42 ns    471578947
bm<uint16_t, Op::Count>/8021/3056        136 ns          105 ns     10000000
bm<uint16_t, Op::Count>/63/62           4.85 ns         3.40 ns    358400000
bm<uint16_t, Op::Count>/31/30           4.33 ns         3.31 ns    471578947
bm<uint16_t, Op::Count>/15/14           5.08 ns         3.71 ns    358400000
bm<uint16_t, Op::Count>/7/6             3.64 ns         2.49 ns    426666667
bm<uint32_t, Op::Count>/8021/3056        254 ns          206 ns      6892308
bm<uint32_t, Op::Count>/63/62           4.60 ns         2.52 ns    670690059
bm<uint32_t, Op::Count>/31/30           4.23 ns         1.80 ns    995555556
bm<uint32_t, Op::Count>/15/14           3.72 ns         1.66 ns   1000000000
bm<uint32_t, Op::Count>/7/6             3.88 ns         1.48 ns    896000000
bm<uint64_t, Op::Count>/8021/3056        724 ns          392 ns      3584000
bm<uint64_t, Op::Count>/63/62           6.15 ns         2.42 ns    426666667
bm<uint64_t, Op::Count>/31/30           4.09 ns         2.07 ns    828499485
bm<uint64_t, Op::Count>/15/14           3.43 ns         1.42 ns   1000000000
bm<uint64_t, Op::Count>/7/6             3.15 ns         1.25 ns    814545455
```