@AlexGuteniev 2024-04-25
> Please meow if you're concerned - I'm willing to accept this hit because I think that the resulting structure is much easier to reason about

I don't like this, I have some respect to small things with tail.
But I've got your concern.

Let me see if I'll be able to optimize this in easier way to reason about, or if making tail scalar would be better here.