@AlexGuteniev 2024-04-25
> Let me see if I'll be able to optimize this in easier way to reason about, or if making tail scalar would be better here.

I think it is good to use `__popcnt` approach for the tail. I've pushed this change.