@StephanTLavavej 2024-04-25
Thanks! :heart_eyes_cat: I pushed a commit to further clarify the comments. Things make so much more sense now!

I verified that my manual 20 GB test case passes. I reran the benchmarks, and it looks like you've reversed the perf damage I did, which is great. Curiously, `bm<uint64_t, Op::Count>/8021/3056` is now 2x faster compared to `main`, which wasn't present in the previous iterations - I'm not sure what changed for the 8-byte codepath though!

Benchmark (Times in ns)             | main | Alex1 | STL  | Alex2
------------------------------------|------|-------|------|-----
`bm<uint8_t, Op::Count>/8021/3056`  |  113 | 62.3  | 62.7 | 61.2
`bm<uint8_t, Op::Count>/63/62`      | 4.70 | 5.36  | 5.80 | 5.56
`bm<uint8_t, Op::Count>/31/30`      | 6.84 | 10.5  | 10.5 | 9.73
`bm<uint8_t, Op::Count>/15/14`      | 6.80 | 9.99  | 10.0 | 6.85
`bm<uint8_t, Op::Count>/7/6`        | 5.10 | 6.62  | 6.59 | 5.11
`bm<uint16_t, Op::Count>/8021/3056` |  277 |  117  |  118 |  116
`bm<uint16_t, Op::Count>/63/62`     | 4.91 | 6.46  | 8.10 | 6.77
`bm<uint16_t, Op::Count>/31/30`     | 4.28 | 6.07  | 7.68 | 6.31
`bm<uint16_t, Op::Count>/15/14`     | 5.13 | 7.21  | 7.06 | 6.78
`bm<uint16_t, Op::Count>/7/6`       | 4.87 | 6.65  | 6.63 | 5.52
`bm<uint32_t, Op::Count>/8021/3056` |  543 |  223  |  224 |  221
`bm<uint32_t, Op::Count>/63/62`     | 5.34 | 6.64  | 7.68 | 6.59
`bm<uint32_t, Op::Count>/31/30`     | 4.49 | 5.78  | 6.84 | 5.77
`bm<uint32_t, Op::Count>/15/14`     | 3.84 | 5.56  | 6.63 | 5.50
`bm<uint32_t, Op::Count>/7/6`       | 4.90 | 5.77  | 5.77 | 5.54
`bm<uint64_t, Op::Count>/8021/3056` |  864 |  872  |  865 |  438 :green_heart:
`bm<uint64_t, Op::Count>/63/62`     | 9.61 | 9.65  | 9.63 | 7.27
`bm<uint64_t, Op::Count>/31/30`     | 6.19 | 6.41  | 6.22 | 5.56
`bm<uint64_t, Op::Count>/15/14`     | 4.70 | 4.61  | 4.93 | 4.67
`bm<uint64_t, Op::Count>/7/6`       | 4.08 | 4.12  | 4.50 | 4.48