@StephanTLavavej 2024-04-26
Thanks again for continuing to refine this important algorithm, and working so hard to educate my cat-sized brain about SIMD! :cat2: :brain: :green_heart: